consolidate all the share dialog cases
