+++ src/window_main/renderer-util.js
@@ -26,7 +26,6 @@ import {
   createSpan,
   queryElements as $$
 } from "../shared/dom-fns";
-import { createSelect } from "../shared/createSelect";
 import * as deckDrawer from "../shared/deck-drawer";
 import { cardType } from "../shared/card-types";
 import { addCardHover } from "../shared/card-hover";
@@ -41,6 +40,7 @@ import {
   toMMSS,
   openScryfallCard
 } from "../shared/util";
+import createShareDialog from './createShareDialog';
 
 const DEFAULT_BACKGROUND = "../images/Bedevil-Art.jpg";
 
@@ -667,7 +667,11 @@ function showColorpicker(
   pickerWrapper.style.boxShadow = "none";
 }
 
-function showDatepicker(defaultDate, onChange = () => {}, pickerOptions = {}) {
+function showDatepicker(
+  defaultDate,
+  onChange = date => {},
+  pickerOptions = {}
+) {
   const cont = createDiv(["dialog_content"]);
   cont.style.width = "320px";
   cont.style.heigh = "400px";
@@ -974,35 +978,7 @@ function createReplayShareButton(draft) {
   const replayShareButton = createDiv(["list_draft_share", draft.id + "dr"]);
   replayShareButton.addEventListener("click", e => {
     e.stopPropagation();
-    const cont = createDiv(["dialog_content"]);
-    cont.style.width = "500px";
-
-    cont.append(createDiv(["share_title"], "Link for sharing:"));
-    const icd = createDiv(["share_input_container"]);
-    const linkInput = createInput([], "", {
-      id: "share_input",
-      autocomplete: "off"
-    });
-    linkInput.addEventListener("click", () => linkInput.select());
-    icd.appendChild(linkInput);
-    const but = createDiv(["button_simple"], "Copy");
-    but.addEventListener("click", function() {
-      ipcSend("set_clipboard", byId("share_input").value);
-    });
-    icd.appendChild(but);
-    cont.appendChild(icd);
-
-    cont.appendChild(createDiv(["share_subtitle"], "<i>Expires in: </i>"));
-    createSelect(
-      cont,
-      ["One day", "One week", "One month", "Never"],
-      "",
-      () => draftShareLink(draft.id, draft),
-      "expire_select"
-    );
-
-    openDialog(cont);
-    draftShareLink(draft.id, draft);
+    createShareDialog(() => draftShareLink(draft.id, draft));
   });
   return replayShareButton;
 }
